+++ api/auth/__init__.py
@@ -1,0 +1,9 @@
+"""Authentication controller of the
+
+"""
+
+from flask import Blueprint
+
+auth = Blueprint('auth', __name__)
+
+import api.auth.routes
